src/main/java/edu/ucsd/sbrg/parameters/ParametersParser.java
package edu.ucsd.sbrg.parameters;

import com.fasterxml.jackson.databind.DeserializationFeature;
import com.fasterxml.jackson.databind.ObjectMapper;

import java.io.IOException;
import java.io.InputStream;

public class ParametersParser {

    public Parameters parse(InputStream parametersJson) throws IOException {

        ObjectMapper mapper = new ObjectMapper();
        mapper.configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);

        return mapper.readValue(parametersJson, Parameters.class);
    }
}
